Config/requirements.txt
# IND-Diplomat Dependencies (Complete Production Stack)

# Core Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
neo4j>=5.0.0
chromadb>=0.4.0
asyncpg>=0.29.0  # PostgreSQL async driver

# LLM Integration
litellm>=1.0.0
ollama>=0.1.0

# ML/NLP
sentence-transformers>=2.2.0
spacy>=3.5.0
numpy>=1.24.0
torch>=2.0.0

# Caching & Session
redis>=4.5.0

# Security & Auth
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0
cryptography>=41.0.0

# Rate Limiting
slowapi>=0.1.8

# Observability
prometheus-client>=0.17.0
structlog>=23.0.0

# Report Generation
reportlab>=4.0.0
python-docx>=0.8.0

# Fuzzy Matching
rapidfuzz>=3.0.0

# UI
streamlit>=1.28.0
plotly>=5.0.0
networkx>=3.0.0
pyvis>=0.3.0

# HTTP Client
aiohttp>=3.8.0
requests>=2.28.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Code Quality (dev)
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
